Add a GitHub action to sync data to S3 (#11)
* Add a GitHub action to sync data to S3

First pass at an "official" version of the cloud sync workflow.
This version uses information from the admin config to get the
name of the S3 bucket and is safe to use on hubs that are not
yet cloud enabled.

* Incorporate schema updates

Based on a conversation in the PR for the cloud admin.json updates,
we decided to rename some cloud.host properties. This commit updates
the GitHub action to reflect those changes.

* Rename the action for clarity

Based on PR feedback, this commit renames the action:
1. to make it clear that it's AWS based
2. to make it clear that it operates on the Hubverse AWS account

* Also sync optional hub directories

Per the PR convo, sync optional hub directories (e.g., target-data)
as well as the required directories.

Additionally, add a note that specific directories can be excluded
by removing them from the "hub_directories" list.