Use rclone for AWS S3 file sync
Resolves Infectious-Disease-Modeling-Hubs/hubverse-cloud#36

Use rclone for cloud syncing because it has a checksum option for
detecting file differences (whereas AWS SYNC relies on update date
and file size, which doesn't work well in a CI environment).

This changeset also ensures that model-outputs are synced to multiple
places in the target S3 bucket: model-outputs/ and raw/model-outputs.
Once the parquet conversion feature is implemented, raw/model-outputs
will reflect the data in a hub's repo, and model-outputs will be the
"user-facing" parquet files.